tests/wpt/web-platform-tests/focus/hasfocus-same-site.html

<!doctype html>
<meta charset=utf-8>
<title>document.hasFocus() with focus in an iframe</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
setup({explicit_done:true});
window.onmessage = function(e) {
    test(function() {
        assert_equals(e.data, "PASS", 'Check result');
    }, "Check result");
    w.close();
    done();
};
var w = window.open("support/hasfocus-same-site-outer.html");
</script>
