third_party/blink/web_tests/external/wpt/focus/hasfocus-same-site-expected.txt

This is a testharness.js-based test.
[FAIL] Check result
  assert_equals: Check result expected "PASS" but got "FAIL: did not have initial focus"
Harness: the test ran to completion.
